backend/requirements.txt
# Core Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.29.0

# Supabase Integration
supabase==2.3.4
gotrue==2.4.2
postgrest==0.13.2

# Geospatial Support
geoalchemy2==0.14.2
shapely==2.0.2
geojson==3.1.0

# NASA GIBS and Earth Data APIs
aiohttp==3.9.1
requests==2.31.0
httpx==0.25.2
xmltodict==0.13.0
Pillow==10.1.0

# AI/ML
tensorflow==2.15.0
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.3

# SMS/USSD
africastalking==1.2.7
twilio==8.10.3

# Caching
redis==5.0.1

# Task Queue
celery==5.3.4

# Environment & Config
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# CORS
fastapi-cors==0.0.6

# Rate Limiting
slowapi==0.1.9

# Logging
loguru==0.7.2
